project/requirements.txt
django
django-tinymce
fontawesomefree
django-crispy-forms
django-recaptcha
pillow
six
psycopg2-binary
django-allauth
gunicorn
